+++ demo/components/playground/playgroundArea.vue
@@ -59,9 +59,9 @@ watch(showMC, val => {
           <LightModeSwitch />
         </div>
         <LayoutAttrs v-model="data.layout" />
-        <div class="text-xs mx-auto text-center mt-8">
+        <div class="mx-auto mt-8 text-center text-xs">
           {{ t('labels.proConfig') }}
-          <a class="block mt-1" target="_blank" rel="noopener" :href="'https://add-to-calendar-pro.com' + (locale !== 'en' ? '/' + locale : '')">
+          <a class="mt-1 block" target="_blank" rel="noopener" :href="'https://add-to-calendar-pro.com' + (locale !== 'en' ? '/' + locale : '')">
             {{ t('labels.discoverPro') }}
             <ArrowTopRightOnSquareIcon class="-mt-0.5 mr-0.5 inline-block h-3 w-3" aria-hidden="true" />
           </a>