+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - v2.5 : PRO option and more
   - ✨ [introducing PRO offering](https://add-to-calendar-pro.com/)
   - ⚠️ Breaking: default branding (check hideBranding option for disabling it)
-  - ⚠️ Breaking: if you want to customize calendar option labels, you need need to do this via the customLabels option 
+  - ⚠️ Breaking: if you want to customize calendar option labels, you need need to do this via the customLabels option
   - new optionsMobile und optionsIOS options to specify device-specific calendar types
   - more css var options and subtle style fixes
   - bundle-size optimization option